+++ packages/happner-2/test/__fixtures/test/integration/security/authentication/workingAuth.js
@@ -1,4 +1,6 @@
-module.exports = class TestAuthProvider extends require('happn-3').providers.SecurityBaseAuthProvider {
+module.exports = class TestAuthProvider extends (
+  require('happn-3').providers.SecurityBaseAuthProvider
+) {
   constructor(securityFacade, config) {
     super(securityFacade, config);
   }
@@ -7,19 +9,28 @@ module.exports = class TestAuthProvider extends require('happn-3').providers.Sec
     return new TestAuthProvider(securityFacade, config);
   }
 
-  async providerCredsLogin(credentials, sessionId) {      
-    if (credentials.username === "secondTestuser@somewhere.com" && credentials.password === "secondPass") {        
-      let user = {username: "secondTestuser@somewhere.com",  groups: { _MESH_GST: { data: {} } }}
+  async providerCredsLogin(credentials, sessionId) {
+    if (
+      credentials.username === 'secondTestuser@somewhere.com' &&
+      credentials.password === 'secondPass'
+    ) {
+      let user = { username: 'secondTestuser@somewhere.com', groups: { _MESH_GST: { data: {} } } };
+      return this.loginOK(credentials, user, sessionId, null, {
+        myToken: 'SecretTokenValueFromAuthProvider',
+      });
+    }
+    return this.loginFailed(credentials.username, 'Invalid credentials');
+  }
+
+  async providerTokenLogin(credentials, decodedToken, sessionId) {
+    if (decodedToken.username === 'secondTestuser@somewhere.com' && credentials.token != null) {
+      let user = { username: 'secondTestuser@somewhere.com', groups: { _MESH_GST: { data: {} } } };
       return this.loginOK(credentials, user, sessionId);
     }
     return this.loginFailed(credentials.username, 'Invalid credentials');
   }
 
-  async providerTokenLogin(credentials, decodedToken, sessionId) {      
-    if (decodedToken.username === "secondTestuser@somewhere.com" && credentials.token != null) {        
-      let user = {username: "secondTestuser@somewhere.com",  groups: { _MESH_GST: { data: {} } }}
-      return this.loginOK(credentials, user, sessionId);
-    }
-    return this.loginFailed(credentials.username, 'Invalid credentials');
+  async providerChangePassword(user, passwordDetails, decodedToken) {
+    throw new Error(`${JSON.stringify(decodedToken)}`);
   }
 };