+++ RELEASES.md
@@ -3,7 +3,15 @@
   - happn-cluster membership updated to mongo-based system
   - tests now use ports auto-assigned by OS
 
-13.0.1 2023-07-21
+3.0.3 2023-08-05
+-----------------
+  - happn-3: auth provider changes
+
+3.0.2 2023-07-25
+-----------------
+  - happner-2: MESH_GST permissions updated not to have {{id}} in path
+
+3.0.1 2023-07-21
 -----------------
   - happner-2: response subscription filter fix