+++ packages/happn-cluster/RELEASES.md
@@ -337,6 +337,11 @@ non-release 2017-12-14
   - fix: happn-3 resetpassword
   - feat: happner-2 resetpassword
 
+12.4.2 2023-08-05
+-----------------
+- feat: fixed mesh_gst paths in happner-2
+
 13.0.0 2023-03-04
 -----------------
   - Mongo-based cluster membership - dependencies update
+